+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Adds unsafe, ID-based API (#80, #81, #82)
 - Adds several functions and methods required [ark-serde](https://github.com/mlange-42/ark-serde) (#83, #85)
 - Adds resource shortcut functions `GetResource` and `AddResource` (#86)
+- Adds `World.Reset` (#89)
 
 ## [[v0.1.0]](https://github.com/mlange-42/ark/tree/v0.1.0)